requirements.txt
pandas==2.2.3
openai==1.51.1
moviepy==1.0.3
Flask==3.0.3
pydub==0.25.1
chromadb==0.5.11
langchain-community==0.3.1
langchain-openai==0.2.2
